MicroserviceVersion/services/musics/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23
python-dotenv==1.0.0
PyJWT==2.8.0

supabase==1.0.3
python-json-logger==2.0.7